refactor(agent): extract ToolCallRow into its own module
AgentPanel.tsx had grown past the 700-line module-size ceiling after the tool-row work. Move ToolCallRow (plus its category-icon map and tone-class helper, and the ~19 icon imports) into ToolCallRow.tsx, dropping AgentPanel.tsx from 737 to 604 lines. Behaviour-preserving.
